requirements.txt
selenium==4.7.2
loguru==0.6.0
python-dotenv==0.21.0